event.kill add tick

+++ pkg/demoinfocs/events/events.go
@@ -140,7 +140,7 @@ type Kill struct {
 	Assister          *common.Player
 	PenetratedObjects int
 	IsHeadshot        bool
-	Tick		  int64
+	Tick              int
 }
 
 // IsWallBang returns true if PenetratedObjects is larger than 0.
